fixed views for creating new sale

+++ sales/api/sales_rest/models.py
@@ -14,7 +14,7 @@ class Customer(models.Model):
     first_name = models.CharField(max_length=100)
     last_name = models.CharField(max_length=100)
     address = models.TextField()
-    phone_number = models.PositiveIntegerField()
+    phone_number = models.BigIntegerField()
     def __str__(self):
         return f"{self.first_name} {self.last_name}"
